moved all dealership stylesheets into new styles directory
