+++ src/components/dealership/card.js
@@ -1,5 +1,6 @@
 import React from "react";
-import "./card.css"
+// import "./card.css"
+import "../../styles/dealerships/card.css"
 
 const DealershipCard = props => {
